user-modify/deprecated-webkit-user-modify-01.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>-webkit-user-modify should be unsupported in author origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#modified {
    -webkit-user-modify: read-write-plaintext-only;
}
</style>
<div id="inherited"></div>
<div id="modified"></div>
<script>
    test(function() {
        const inherited = document.getElementById('inherited');
        assert_equals(inherited.style.WebkitUserModify, undefined);
        assert_equals(inherited.style.webkitUserModify, undefined);
        assert_equals(getComputedStyle(inherited).getPropertyValue('-webkit-user-modify'), '');
    }, '-webkit-user-modify is not supported');

    test(function() {
        const modified = document.getElementById('modified');
        assert_equals(getComputedStyle(modified).getPropertyValue('-webkit-user-modify'), '');
        assert_equals(modified.style.WebkitUserModify, undefined);
        assert_equals(modified.style.webkitUserModify, undefined);
        assert_not_equals(getComputedStyle(modified).getPropertyValue('-webkit-user-modify'), 'read-write-plaintext-only');
    }, '-webkit-user-modify is not supported in author origin CSS');

    test(function() {
        const inherited = document.getElementById('inherited');
        const modified = document.getElementById('modified');
        assert_equals(getComputedStyle(modified).getPropertyValue('-webkit-user-modify'), getComputedStyle(inherited).getPropertyValue('-webkit-user-modify'));
    }, '-webkit-user-modify remains as empty string');
</script>
